requirements.txt
flask
psycopg2-binary
requests
beautifulsoup4
python-dotenv
